% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.distributions.R
\docType{data}
\name{get.copula}
\alias{get.copula}
\title{get.copula}
\format{
An object of class \code{list} of length 5.
}
\usage{
get.copula
}
\value{
A list of copula.
}
\description{
Storing list of copula.
}
\keyword{datasets}
