% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.distributions.R
\docType{data}
\name{get.distributions}
\alias{get.distributions}
\title{get.distributions}
\format{
An object of class \code{list} of length 9.
}
\usage{
get.distributions
}
\value{
A list of distributions.
}
\description{
Storing list of distributions for Biomarker, X and Time-to-event, T.
}
\examples{
get.distributions
#> "exponential" "weibull"     "gaussian"    "normal"      "lognormal"
#> "gompertz"    "skewnormal"
}
\keyword{datasets}
