% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeroc_gof.R
\name{nelson_aalen}
\alias{nelson_aalen}
\title{nelson_aalen}
\usage{
nelson_aalen(df)
}
\arguments{
\item{df}{A data frame with column t = time-to-event, event = status(0 or 1)}
}
\value{
A data frame of survival estimate.
}
\description{
Function to estimate survival probability using the Nelson-Aalen estimator.
}
\keyword{internal}
