% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeroc_predict.R
\name{pars.boot}
\alias{pars.boot}
\title{pars.boot}
\usage{
pars.boot(X, boot.val, pargs)
}
\arguments{
\item{X}{A sequence of integer specifying the cutoff to be used.\cr}

\item{boot.val}{An integer specifying the iteration for the bootstrap process.\cr}

\item{pargs}{A list of named vector specifying the parameters for X, T and PH/copula.\cr}
}
\value{
A list of random parameters.
}
\description{
helper function to generate random parameters used in bootstrap process.
}
\keyword{internal}
