% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.fitTROC}
\alias{print.fitTROC}
\title{print.fitTROC}
\usage{
\method{print}{fitTROC}(x, ...)
}
\arguments{
\item{x}{A fitTROC object}

\item{...}{Additional argument (not use)}
}
\value{
Summarize estimated parameters in console
}
\description{
print.fitTROC
}
\examples{
test <- timeroc_obj(dist = 'lognormal-weibull-PH')
rr <- rtimeroc(test, n=300, params.x=c(meanlog=1,sdlog=0.8),
               params.t = c(shape=1.6,scale=1.2),
               params.ph = 1.1)
cc <- timeroc_fit(test, rr$x, rr$t, rr$event)
print(cc)

}
