% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.TimeROC}
\alias{print.TimeROC}
\title{print.TimeROC}
\usage{
\method{print}{TimeROC}(x, ...)
}
\arguments{
\item{x}{A TimeROC object}

\item{...}{Additional arguments (not use)}
}
\value{
Summarize model's info in console
}
\description{
print.TimeROC
}
\examples{
test <- timeroc_obj(dist = 'lognormal-weibull-PH')
print(test)

}
