% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeroc_predict.R
\name{roc.cop}
\alias{roc.cop}
\title{roc.cop}
\usage{
roc.cop(x, t, xfn, tfn, cfn, pars, ci)
}
\arguments{
\item{x}{An integer vector specifying the biomarker value.\cr}

\item{t}{An integer vector specifying the time to produce the ROC curve.\cr}

\item{xfn}{A list of functions used to compute the biomarker distribution.\cr}

\item{tfn}{A list of functions used to compute the time-to-event distribution.\cr}

\item{cfn}{A list of functions used to compute the copula distribution.\cr}

\item{pars}{A list of vector specifying the estimated X, T and PH/copula parameter.\cr}

\item{ci}{An integer specifying the confidence interval of the ROC curve.\cr}
}
\description{
Function that will be automatically called within troc_curve & timeroc to compute sensitivity and specificity of a copula mdoel.
}
\keyword{internal}
