% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeroc_predict.R
\name{timeroc}
\alias{timeroc}
\title{timeroc}
\usage{
timeroc(model, pars, t, cutoff, x.dist, t.dist, copula, ci)
}
\arguments{
\item{model}{A string of either PH/copula.\cr}

\item{pars}{A list of X, T and PH/copula parameters.\cr}

\item{t}{A vector of integer specifying the time to produce the ROC curve.\cr}

\item{cutoff}{A vector of integer specifying the biomarker value needed for sensitivity & specificity computation.\cr}

\item{x.dist}{A string specifying the biomarker distribution.\cr}

\item{t.dist}{A string specifying the time-to-event distribution.\cr}

\item{copula}{A string specifying the type of copula.\cr}

\item{ci}{An integer specifying the confidence interval for the estimated ROC curve.\cr}
}
\description{
helper function that automatically called within troc_curve function.
}
\keyword{internal}
