% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeroc_obj.R
\name{timeroc_obj}
\alias{timeroc_obj}
\title{timeroc_obj}
\usage{
timeroc_obj(
  dist,
  params.x = NA,
  params.t = NA,
  copula = NA,
  params.copula = NA,
  params.ph = NA
)
}
\arguments{
\item{dist}{A string emphasizing the distribution assumption for biomarker-time-model.\cr}

\item{params.x}{Vector of biomarker parameter.\cr}

\item{params.t}{Vector of time-to-event parameter.\cr}

\item{copula}{A string emphasizing on the copula to be used.\cr}

\item{params.copula}{An integer for copula parameter.\cr}

\item{params.ph}{An integer for association parameter.}
}
\value{
A 'TimeROC' object.
}
\description{
Function to initialized time-dependent ROC object.
}
\examples{
## Copula model
test <- timeroc_obj(dist = 'gompertz-gompertz-copula', copula = "gumbel90")

## PH model
test <- timeroc_obj(dist = 'weibull-gompertz-PH')
}
