% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/timeroc_predict.R
\name{troc.curve}
\alias{troc.curve}
\title{troc.curve}
\usage{
troc.curve(pars, cutoff, t, x.dist, t.dist, copula, ci)
}
\arguments{
\item{pars}{A list of vector specifying the parameters for X, T and PH/copula.\cr}

\item{cutoff}{A vector of integer specifying the biomarker value to be used in the ROC computation.\cr}

\item{t}{A vector of integer specifying the time to produce the ROC curve.\cr}

\item{x.dist}{A string to specify the biomarker distribution.\cr}

\item{t.dist}{A string to specify the time-to-event distribution.\cr}

\item{copula}{A string to specify the copula distribution.\cr}

\item{ci}{An integer to specify the confidence interval for the estimated ROC curve.\cr}
}
\value{
A list of dataframe.
}
\description{
helper function to estimate sensitivity and specificity.
}
\keyword{internal}
